Decompose API handlers by resource
